# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Saveur API
  version: 0.1.0
paths:
  /api/recipes:
    get:
      operationId: listRecipes
      summary: List recipes
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search by title, ingredient, or tag.
      responses:
        "200":
          description: Recipe list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeListResponse"
    post:
      operationId: createRecipe
      summary: Create a manual recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaseRecipe"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/recipes/{id}:
    parameters:
      - $ref: "#/components/parameters/RecipeIdParam"
    get:
      operationId: getRecipe
      summary: Get recipe by id
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRecipe
      summary: Update recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaseRecipe"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRecipe
      summary: Delete recipe
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteRecipeResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/import/url:
    post:
      operationId: importRecipeFromUrl
      summary: Import a recipe from URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportUrlRequest"
      responses:
        "200":
          description: Import succeeded or partially succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Import failed with unusable output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/recipes/{id}/reimport:
    parameters:
      - $ref: "#/components/parameters/RecipeIdParam"
    post:
      operationId: reimportRecipe
      summary: Reimport an existing URL recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReimportRecipeRequest"
      responses:
        "200":
          description: Reimport succeeded or partially succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Only URL recipes may be reimported or import failed
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ImportResult"
                  - $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/recipes/{id}/photo:
    parameters:
      - $ref: "#/components/parameters/RecipeIdParam"
    post:
      operationId: uploadRecipePhoto
      summary: Upload recipe hero photo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Recipe updated with uploaded hero photo URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          description: Validation failed or object storage unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    SourceType:
      type: string
      enum:
        - URL
        - SCAN
        - MANUAL
    Unit:
      type: string
      enum:
        - UNIT
        - ML
        - GRAM
        - KG
        - TSP
        - TBSP
        - PINCH
        - HANDFUL
        - UNKNOWN
    ImportStatus:
      type: string
      enum:
        - SUCCESS
        - PARTIAL
        - FAILED
    IngredientData:
      type: object
      description: A single ingredient with normalized fields.  Quantity fields should be filled where possible, but quantityText should always be filled to preserve the original text and any nuances that can't be captured in structured fields.
      additionalProperties: false
      required:
        - name
        - quantityText
        - quantityValue
        - quantityMin
        - quantityMax
        - unit
        - isWholeItem
        - optional
        - isPantryItem
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Ingredient name without quantity.  Sentence case.
          examples:
            - Curry powder
            - White potatoes
            - Eggs
            - Whole milk
        quantityText:
          type: [ string, "null" ]
          maxLength: 40
          description: Original quantity text, preserving fractions/ranges/to taste.
          examples:
            - "1tsp"
            - "2 cups"
            - "1/2"
            - "1-2"
            - "to taste"
            - "a few sprigs"
        quantityValue:
          type: [ number, "null" ]
          description: Primary numeric quantity value when parseable.
          examples: [ 1, 0.5 ]
        quantityMin:
          type: [ number, "null" ]
          description: Minimum quantity, when the quantity text represents a range.
          examples: [ 1 ]
        quantityMax:
          type: [ number, "null" ]
          description: Maximum quantity, when the quantity text represents a range.  Must be larger than quantityMin.
          examples: [ 2 ]
        unit:
          type: string
          description: Unit of measurement for the quantity. Use UNKNOWN when unit is unknown or not explicit. Otherwise choose from allowed enum values.  Convert where needed to fit these units (for example, convert 1000g to 1kg).
          enum:
            - UNIT
            - ML
            - GRAM
            - KG
            - TSP
            - TBSP
            - PINCH
            - HANDFUL
            - UNKNOWN
          examples: [ GRAM, TSP, UNKNOWN ]
        isWholeItem:
          type: boolean
          description: True for countable whole items that are typically indivisible (eggs, garlic cloves), false for non-countable items (flour, milk, sugar) or ingredients that are typically measured by weight or volume (onions, potatoes).
        optional:
          type: boolean
          description: True only if ingredient is optional, such as a garnish, add-in or optional flavouring.
        isPantryItem:
          type: boolean
          description: True for commonly assumed pantry staples (for example salt, pepper, oil, water) that can be assumed available in most kitchens.
    PrepTaskData:
      type: object
      additionalProperties: false
      description: A single prep task, which uses one or more raw ingredients and should produce one single prepared ingredient that can be referenced in cook steps.  Prep tasks are for maximizing mise en place, creating prepared ingredients that may be referenced in cook steps. Chopping, slicing, washing, spice mixes, stocks, etc. are prep tasks.  Prep tasks should be used when there is a clear transformation or when multiple raw ingredients are combined together, even if heat is involved, as long as the task can be done ahead of time and isn't time-sensitive.  For example, "chop onion and tomatoes and combine with spices" would be a prep task that produces a prepared ingredient like "spice mix" or "diced tomato", which is then added in a later cook step.
      required:
        - preparationName
        - detail
        - sourceIngredients
      properties:
        preparationName:
          type: string
          minLength: 3
          maxLength: 30
          description: Name of the resulting prepared ingredient name.  Must not conflict with any raw ingredient names (IngredientData.name).  Where possible use a concise name that captures the key transformation ('diced tomatoes') or intended purpose of the prepared ingredient ('marinade').  This name should be used in the sourceIngredients list of cook steps to refer to this prepared ingredient.  Max 30 chars.
          examples:
            - Spice mix
            - Stock
            - Soup base
            - Marinade
            - Chopped onion
            - Prepared vegetables
            - Sauce
            - Carrot mixture
        sourceIngredients:
          type: array
          minItems: 1
          maxItems: 20
          description: List of ingredient names used in this prep step.  Must match ingredient names from the ingredient list (IngredientData.name).
          items:
            type: string
            minLength: 1
            maxLength: 120
            examples:
              - Curry powder
              - Turmeric
              - Eggs
              - Onions
              - Tomatoes
        detail:
          type: [ string, "null" ]
          maxLength: 1200
          description: Detailed instructions for the prep task. Include all processing details and ingredient names.  Supports markdown.  Put raw ingredient names in **bold** but don't bold names of pantry items like salt, pepper, oil or water.
          examples:
            - Combine the **curry powder** and **turmeric** in a small bowl
            - Mix the **lemongrass** and **lime leaf** paste, **gluten free soy sauce**, **honey**, and **dried chilli flakes** together to make the marinade
            - Chop the **onion** and **tomatoes** and combine with a pinch of salt to make the base for the sauce
            - Wash and chop the **cilantro** and **mint** to make the herb garnish
    CookStepData:
      type: object
      description: Ordered, heat/time-sensitive cooking step that cannot be done ahead of time.
      additionalProperties: false
      required:
        - instruction
        - detail
        - sourceIngredients
        - timerSeconds
      properties:
        instruction:
          type: string
          minLength: 1
          maxLength: 50
          description: Terse description of the cooking step, ideally starting with a verb. Focus on one primary action, even if multiple things are happening in the step, and put details in the detail field.  For example, say "Add spices" instead of "Add spices and cook".
          examples:
            - Fry the onion
            - Bake fish
            - Add sauce and simmer
            - Cook rice
            - Add spice mix
            - Garnish and serve
        detail:
          type: [ string, "null" ]
          maxLength: 1400
          description: Detailed cooking instructions, including sequencing and timing details.  Include all relevant ingredient names.  Supports markdown.  Put ingredient names (both raw and prepared) in **bold** but don't bold names of pantry items like salt, pepper, oil or water.
          examples:
            - Add the **diced tomato** with a pinch of salt and sugar and cook for 3-4 min or until the tomatoes have broken down
        sourceIngredients:
          type: array
          minItems: 1
          maxItems: 20
          description: List of raw ingredient and prepared ingredient names used in this cook step.  Must match ingredient names from the ingredient list (IngredientData.name) or prep task names (PrepTaskData.preparationName).
          items:
            type: string
            minLength: 1
            maxLength: 120
            examples:
              - Curry powder
              - Eggs
              - Tomatoes
              - Spice mix
              - Peeled eggs
              - Diced tomatoes
        timerSeconds:
          type: [integer, "null"]
          minimum: 0
          maximum: 86400
          description: Timer duration in seconds for timed steps. If no duration is explicitly mentioned, set null.
          examples: [ 360, null ]
    ImportConfidence:
      type: object
      additionalProperties: false
      required:
        - title
        - ingredients
        - cookSteps
        - overall
      description: Model confidence scores from 0-1.
      properties:
        title:
          type: number
          minimum: 0
          maximum: 1
          examples:
            - 0.98
        ingredients:
          type: number
          minimum: 0
          maximum: 1
          examples:
            - 0.9
        cookSteps:
          type: number
          minimum: 0
          maximum: 1
          examples:
            - 0.9
        overall:
          type: number
          minimum: 0
          maximum: 1
          examples:
            - 0.92
    RecipeMetaData:
      type: object
      description: Top-level, user-editable recipe metadata fields that are not specific to ingredients or cooking steps. 
      additionalProperties: false
      required:
        - title
        - description
        - heroPhotoUrl
        - servingCount
        - timeRequiredMinutes
        - tags
        - sourceType
        - sourceRef
        - importPrompt
        - importRunId
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 140
          description: Human-friendly recipe title.
          examples:
            - Classic Kedgeree With Roasted Tomatoes
        description:
          type: [ string, "null" ]
          maxLength: 1200
          description: Short summary of the dish or recipe description from source.
          examples:
            - Rich and comforting kedgeree with smoky fish and roasted tomatoes.
        heroPhotoUrl:
          type: [ string, "null" ]
          description: URL of the best available hero image of the dish, likely the largest image on the page.
          examples:
            - https://example.com/recipe-photo.jpg
        servingCount:
          type: [ integer, "null" ]
          minimum: 1
          maximum: 20
          description: Default number of servings this recipe is written for.
          examples:
            - 2
        timeRequiredMinutes:
          type: [ integer, "null" ]
          minimum: 1
          maximum: 600
          description: Total time in minutes when inferable.
          examples:
            - 35
        tags:
          type: array
          maxItems: 24
          items:
            type: string
            minLength: 1
            maxLength: 40
            examples:
              - Fish
              - British
              - Gluten Free
          description: Flat tags inferred from source categories, cuisine, or theme.
        sourceType:
          $ref: "#/components/schemas/SourceType"
        sourceRef:
          type: string
          description: Original source reference, such as the URL for URL-sourced recipes or a note about where a manually entered recipe came from.
        importPrompt:
          type: [ string, "null" ]
          description: An additional prompt used during an LLM import to help guide the model.
          maxLength: 4000
        importRunId:
          type: [ string, "null" ]
          description: The ID of the import run that created this recipe, if applicable.
          format: uuid
    BaseRecipe:
      description: All user-editable fields of a recipe, including metadata, ingredient list, prep tasks and cook steps (excluding system-generated fields like id and timestamps).
      allOf:
        - $ref: "#/components/schemas/RecipeMetaData"
        - type: object
          additionalProperties: false
          required:
            - ingredients
            - prepTasks
            - cookSteps
          properties:
            ingredients:
              type: array
              minItems: 1
              maxItems: 120
              description: Complete raw ingredient list with normalized fields.
              items:
                $ref: "#/components/schemas/IngredientData"
            prepTasks:
              type: array
              maxItems: 80
              description: Tasks to maximize mise en place, creating prepared ingredients that may be referenced in cook steps. Chopping, slicing, washing, spice mixes, stocks, and non-heat-sensitive combinations are prep tasks.
              items:
                $ref: "#/components/schemas/PrepTaskData"
            cookSteps:
              type: array
              minItems: 1
              maxItems: 120
              description: Ordered heat/time-sensitive cooking steps that cannot be done ahead of time.
              items:
                $ref: "#/components/schemas/CookStepData"
    Recipe:
      description: A complete recipe including all system-generated fields.
      allOf:
        - $ref: "#/components/schemas/BaseRecipe"
        - type: object
          additionalProperties: false
          required:
            - id
            - ownerId
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
            ownerId:
              type:
                - string
                - "null"
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    RecipeFromLlmImport:
      description: A recipe created from an LLM import, which includes additional fields related to the import process and model confidence, and narrows the sourceType and sourceRef to be specific to URL imports.
      allOf:
        - $ref: "#/components/schemas/BaseRecipe"
        - type: object
          additionalProperties: false
          required:
            - sourceType
            - sourceRef
          properties:
            sourceType:
              type: string
              const: URL
            sourceRef:
              type: string
              description: Original URL the recipe was imported from
        - type: object
          additionalProperties: false
          required:
            - confidence
            - warnings
          properties:
            confidence:
              $ref: "#/components/schemas/ImportConfidence"
            warnings:
              type: array
              maxItems: 20
              items:
                type: string
                minLength: 1
                maxLength: 220
                examples:
                  - Could not determine a hero image URL
              description: Non-fatal caveats discovered while extracting data.
    ImportUrlRequest:
      type: object
      additionalProperties: false
      required:
        - url
      properties:
        url:
          type: string
        prompt:
          type:
            - string
            - "null"
          maxLength: 4000
    ReimportRecipeRequest:
      type: object
      additionalProperties: false
      properties:
        prompt:
          type:
            - string
            - "null"
          maxLength: 4000
    ImportResult:
      type: object
      additionalProperties: false
      required:
        - recipeId
        - importRunId
        - status
        - usable
        - warnings
        - adapter
      properties:
        recipeId:
          type: string
          format: uuid
        importRunId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ImportStatus"
        usable:
          type: boolean
        warnings:
          type: array
          items:
            type: string
        adapter:
          type: string
    RecipeListResponse:
      type: object
      additionalProperties: false
      required:
        - recipes
      properties:
        recipes:
          type: array
          items:
            $ref: "#/components/schemas/Recipe"
    RecipeResponse:
      type: object
      additionalProperties: false
      required:
        - recipe
      properties:
        recipe:
          $ref: "#/components/schemas/Recipe"
    DeleteRecipeResponse:
      type: object
      additionalProperties: false
      required:
        - deleted
      properties:
        deleted:
          type: boolean
    ValidationIssue:
      type: object
      additionalProperties: true
      properties:
        instancePath:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: true
      required:
        - error
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
        detail:
          type: string
  parameters:
    RecipeIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid